Update README.md
Changed the entirety to English instead of Swedish. Used simple titles instead of explaining code-skeleton, giving examples of usercase-senarios (which were increadibly stupid to begin with), ect. Also listed sources a short explenation on how to run the programme. 

I earlier used this project as a project for a programming course on the engineering-physics programme.